+++ src/actividades/Actividad.css
@@ -1,0 +1,35 @@
+
+  .header {
+    text-align: center;
+    font-size: 30px;
+    font-weight:bold;
+  }
+  
+  body{margin-top:20px;}
+  
+  .comment-wrapper .panel-body {
+      max-height:650px;
+      overflow:auto;
+  }
+  
+  .comment-wrapper .media-list .media img {
+      width:64px;
+      height:64px;
+      border:2px solid #e5e7e8;
+  }
+  
+  .comment-wrapper .media-list .media {
+      border-bottom:1px dashed #efefef;
+      margin-bottom:25px;
+  }
+  
+  .nav {
+    border-style:outset;
+    font-size: large;
+    font-weight: bold;
+  }
+  
+  
+
+
+
